rename library.h into system.h
